requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
dj-database-url==3.0.1
Django==6.0
django-jazzmin==3.0.1
fonttools==4.61.0
gunicorn==23.0.0
holidays==0.86
packaging==25.0
pango==0.0.1
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
reportlab==4.4.5
six==1.17.0
sqlparse==0.5.4
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==67.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
python-dotenv